Refactor TrackFrame similarly to TrackClip
