+++ src/meson.build
@@ -19,6 +19,7 @@ keyframe_sources = [
 	'timeline/tracks/keyframe-track-clip-layout.c',
 	'timeline/tracks/keyframe-track-clip-widget.c',
 	'timeline/tracks/keyframe-track-frame.c',
+	'timeline/tracks/keyframe-track-frame-layout.c',
 	'model/keyframe-layer.c',
 	'model/keyframe-layer-geometry.c',
 	'model/keyframe-layer-text.c',
