+++ src/timeline/tracks/keyframe-track-clip-private.h
@@ -2,13 +2,9 @@
 
 #include <gtk/gtk.h>
 
-#include "keyframe-track.h"
+#include "keyframe-tracks.h"
 
 G_BEGIN_DECLS
-
-#define KEYFRAME_TYPE_TRACK_CLIP (keyframe_track_clip_get_type())
-
-G_DECLARE_FINAL_TYPE (KeyframeTrackClip, keyframe_track_clip, KEYFRAME, TRACK_CLIP, KeyframeTrack)
 
 // TODO: Make Private?
 struct _KeyframeTrackClip
@@ -26,8 +22,6 @@ struct _KeyframeTrackClip
 
     float start_position;
 };
-
-GtkWidget *keyframe_track_clip_new (void);
 
 
 // Layout Manager
